renamed http module to http_api and fixed naming collision
